denormalize message
